Refactor(timer.js): 파일명 getTimeDiff에서 timer로 수정
1. getTimeDiff 함수 로직 세분화. e.g. 01 minutes -> 1 minutes 뜨도록 수정
2. formatDate 함수 하나로 통합 리팩토링
